Refactor and minor bugfixes
