+++ app/js/admin.js
@@ -3,23 +3,24 @@ import React from "react";
 import ReactDOM from "react-dom";
 import RoomView from "./room";
 import styled from "styled-components";
+import { ModalRoot, ModalProvider } from "./modals/modals";
 
 const Users = () => {
   return (
     <div id="users" className="col s12">
-        <ul className="collection">
-            <a href="/user_preferences/" className="collection-item">Preferences</a>
-            <a href="/accounts/mail/" className="collection-item">Email users</a>
-        </ul>
+      <ul className="collection">
+        <a href="/user_preferences/" className="collection-item">Preferences</a>
+        <a href="/accounts/mail/" className="collection-item">Email users</a>
+      </ul>
     </div>
   )
 }
 
 const Blog = () => {
   return (
-    <div id="blog" class="col s12">
-      <ul class="collection">
-        <a href="/blog/create" class="collection-item">Add blog post</a>
+    <div id="blog" className="col s12">
+      <ul className="collection">
+        <a href="/blog/create" className="collection-item">Add blog post</a>
       </ul>
     </div>
   )
@@ -27,12 +28,12 @@ const Blog = () => {
 
 const Zosia = () => {
   return (
-    <div id="zosia" class="col s12">
-      <ul class="collection">
-        <a href="/conferences/" class="collection-item">Conferences</a>
-        <a href="/bus/" class="collection-item">Bus</a>
-        <a href="/accounts/organizations/" class="collection-item">Organizations</a>
-        <a href="/conferences/export_data/" class="collection-item"> Export data</a>
+    <div id="zosia" className="col s12">
+      <ul className="collection">
+        <a href="/conferences/" className="collection-item">Conferences</a>
+        <a href="/bus/" className="collection-item">Bus</a>
+        <a href="/accounts/organizations/" className="collection-item">Organizations</a>
+        <a href="/conferences/export_data/" className="collection-item"> Export data</a>
       </ul>
     </div>
   )
@@ -40,39 +41,39 @@ const Zosia = () => {
 
 const Sponsors = () => {
   return (
-    <div id="sponsors" class="col s12">
-      <ul class="collection">
-          <a href="/sponsors/" class="collection-item">Sponsors</a>
-          <a href="/sponsors/create" class="collection-item">Add sponsor</a>
+    <div id="sponsors" className="col s12">
+      <ul className="collection">
+        <a href="/sponsors/" className="collection-item">Sponsors</a>
+        <a href="/sponsors/create" className="collection-item">Add sponsor</a>
       </ul>
     </div>
   )
 }
 
 const QA = () => (
-  <div id="qa" class="col s12">
-    <ul class="collection">
-      <a href="/questions/all/" class="collection-item">Q&A</a>
-      <a href="/questions/add/" class="collection-item">Q&A add</a>
+  <div id="qa" className="col s12">
+    <ul className="collection">
+      <a href="/questions/all/" className="collection-item">Q&A</a>
+      <a href="/questions/add/" className="collection-item">Q&A add</a>
     </ul>
   </div>
 )
 
 const Lectures = () => (
-  <div id="lectures" class="col s12">
-    <ul class="collection">
-      <a href="/lectures/all" class="collection-item">Lectures</a>
-      <a href="/lectures/create" class="collection-item">Add lecture</a>
-      <a href="/lectures/schedule/update/" class="collection-item">Update schedule</a>
+  <div id="lectures" className="col s12">
+    <ul className="collection">
+      <a href="/lectures/all" className="collection-item">Lectures</a>
+      <a href="/lectures/create" className="collection-item">Add lecture</a>
+      <a href="/lectures/schedule/update/" className="collection-item">Update schedule</a>
     </ul>
   </div>
 )
 
 const Rooms = () => (
-  <div id="rooms" class="col s12">
-    <ul class="collection">
-      <a href="/rooms/report/" class="collection-item">Rooms</a>
-      <a href="/rooms/import/" class="collection-item">Add room</a>
+  <div id="rooms" className="col s12">
+    <ul className="collection">
+      <a href="/rooms/report/" className="collection-item">Rooms</a>
+      <a href="/rooms/import/" className="collection-item">Add room</a>
     </ul>
   </div>
 )
@@ -80,7 +81,7 @@ const Rooms = () => (
 const AdminView = props => {
   const [current_view, setView] = React.useState("rooms");
   const views = {
-    "users": { 
+    "users": {
       component: Users,
       name: "Users"
     },
@@ -104,7 +105,7 @@ const AdminView = props => {
       component: Lectures,
       name: "Schedule"
     },
-    "rooms" : {
+    "rooms": {
       component: RoomView,
       name: "Rooms"
     },
@@ -112,36 +113,37 @@ const AdminView = props => {
   const Component = views[current_view].component;
   return (
     <div className="container">
-        <h3>Admin panel</h3>
-
-        <div className="row">
-            <div className="col s12">
-                <ul className="tabs">
-                {Object.keys(views).map(view_id => 
-                  <li className="tab col" key={view_id}>
-                    <a 
-                      href="#" 
-                      className={view_id === current_view ? "active" : "" }
-                      onClick={() => setView(view_id)}
-                    >
-                      {views[view_id].name}
-                    </a>
-                  </li>)}
-                </ul>
-            </div>
-
+      <h3>Admin panel</h3>
+      <div className="row">
+        <div className="col s12">
+          <ul className="tabs">
+            {Object.keys(views).map(view_id =>
+              <li className="tab col" key={view_id}>
+                <a
+                  href="#"
+                  className={view_id === current_view ? "active" : ""}
+                  onClick={() => setView(view_id)}
+                >
+                  {views[view_id].name}
+                </a>
+              </li>)}
+          </ul>
         </div>
-        <div className="row">
-            <Component></Component>
 
-        </div>
+      </div>
+      <div className="row">
+        <ModalProvider>
+          <ModalRoot />
+          <Component></Component>
+        </ModalProvider>
+      </div>
     </div>
   )
 }
 
 ReactDOM.render(
-    (<AdminView/>),
-    document.getElementById('react-root')
+  (<AdminView />),
+  document.getElementById('react-root')
 );
